Rename `constant_mod`/`runtime_mod` to `*residue`
Renames the following:
- `modular::constant_mod` => `modular::residue`
- `modular::runtime_mod` => `modular::dyn_residue`

Notably these names map to the types they contain, which should make it
a bit more intuitively obvious what's inside of them.

It was also difficult to pick a module name for a prospective
`BoxedResidue`, so now it can be `modular::boxed_residue`.
