+++ src/modular/residue.rs
