+++ src/modular/residue/macros.rs
